PONDERADA-4-PROG/backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
orjson==3.9.7
pydantic==2.4.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
